% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cvl4acp.R
\name{plot.cvl4acp}
\alias{plot.cvl4acp}
\title{plot the cross-validation curve produced by cvl4acp}
\usage{
\method{plot}{cvl4acp}(x, logscale = FALSE, ...)
}
\arguments{
\item{x}{fitted "cvl4acp" object}

\item{logscale}{If it is TRUE, the x-axis is shown by log scale. By default it is FALSE}

\item{...}{Other argumnets that can be passed to cvl4acp}
}
\description{
Plots the cross-validation curve, as a function of the lambda values used.
}
\details{
A plot is produced, and nothing is returned. Red line indicates the selected lambda with minimal tested error. Blue line is for lambda with one standard deviation above, which is \code{lambda.1se} of the fitted model.
}
\examples{
x <- matrix(rnorm(1000), nrow = 50, ncol = 20)
q <- runif(50)
y <- x[,c(1,3)] \%*\% c(1,1) + x[,c(2,3)] \%*\% c(-1,1) * (q < 0.5) + 0.5 * rnorm(50)
cvfit <- cvl4acp(x,y,q, nfolds=5)
plot(cvfit, logscale=TRUE)

}
