% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.l4acp.R
\name{plot.l4acp}
\alias{plot.l4acp}
\title{plot the threshold grid curve produced by l4acp}
\usage{
\method{plot}{l4acp}(x, s = min(x$lambda), type = "lasso", ...)
}
\arguments{
\item{x}{fitted "l4acp" object}

\item{s}{the specified lambda value for threshold grid curve. Note that when specific lambda values are putted, it does not come out with the result from that lambda values exactly. Rather it finds the nearest lambda value from the lambda sequence in the fitted "l4acp" object. By default it is the smallest one from the lambda sequence.}

\item{type}{If the type is "lasso", it makes the plot by the value of a lasso objective function. Else if the type is "mse", it makes the plot by the value of a MSE.}

\item{...}{other arguments that can be passedd to l4acp}
}
\description{
Plots the lasso objective function's value obtained from running regression by putting each value of threshold parameter's range in a model. It can also make plot not from lasso objective function but from ordinary MSE. Since \code{l4acp} can use multiple lambda values, it is required to specify a lambda value. Note that the plot has x-axis corresponding to threshold grid, and y-axis corresponding to the lasso objective function.
}
\details{
A plot is produced, and nothing is returned
}
\examples{
# using 'lr' as an threshold variable
data("grdata")
lrdata <- grdata[(!is.na(grdata[,"lr"])),]
x <- lrdata[,c(4:dim(grdata)[2])]
q <- lrdata[,"lr"]
y <- lrdata[,"gr"]
fit <- l4acp(x=x, y=y, q=q, s=0.00094)
plot(fit, type="lasso")
}
