% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l4acp.R
\name{l4acp}
\alias{l4acp}
\alias{l4acpEst}
\title{l4acp}
\usage{
l4acp(x, y, q, s, trim=0.1)
}
\arguments{
\item{x}{Covariates.}

\item{y}{A dependent variable.}

\item{q}{A threshold covariate.}

\item{s}{The values of lambda used for lasso regression}

\item{trim}{The percentile for trimming the data of a threshold covariate (from above and below) to obtain the range for threshold parameter. By default, it is 0.1. And this results in finding a threshold parameter from 10-90 percentile range of a threshold variable.}
}
\value{
\item{lambda}{ The values of lambda corresponding to the each regression outcome below. It is automatically ordered from high to low values.}
\item{tau.hat}{ The estimated threshold covariates.}
\item{coefficients}{ The regression coefficients for covariates. If the covariates has n different variables, it leads to (2n+2) coefficients.
The first (n+1) coefficients can be interpreted as the coefficients when a threshold covariate is below the threshold parameter. It includes an intercept at the head. That's why the number of coefficients is (n+1), not n.
And the next (n+1) variables are change of regression coefficients when a threshold covariate exceeds the threshold parameter, also including change of an intercept.}
\item{adj.R2}{ Adjusted R2 square.}
\item{M.alpha}{ The number of nonzero coefficients from regression, whose maximum is (2n+2).}
\item{grid.loop}{ Used for the threshold grid curve.}
\item{matrix}{ A matrix that includes \code{lambda}, \code{tau.hat}, \code{coefficients}, \code{M.alpha}, \code{R2}, and \code{adj.R2}}
}
\description{
LASSO regression of a model with a change point due to a covariate threshold. It obtains regression coefficients for covariates, and a threshold parameter.
}
\examples{
# using 'gdp60' as an threshold variable
data("grdata")
x <- grdata[,c(5:dim(grdata)[2])]
q <- grdata[,"gdp60"]
y <- grdata[,"gr"]
fit <- l4acp(x,y,q, s=0.0007)
plot(fit)

}
\references{
Sokbae Lee, Myung Hwan Seo, and Youngki Shin, (2016) \emph{The Lasso for High-Dimensional Regression with a Possible Change Point}, Journal of the Royal Statistical Society Series B, Vol 78(1), 193-210
}
\seealso{
\code{plot.l4acp} and \code{cvl4acp}
}
